Attempt to add links for files included in Kernel require once
coworker is failing cos these files are not in it's path. It's path for some reason includes
drupal/sites/all/modules/civicrm/ but NOT
core

I tried to add it in the Kernel file but it is not picking it up.
I think the issue is that coworker has long running processes and does
not reload the main php files very often - holding loaded files in memory

Once we have fixed the path in coworker these can go

Bug: T400419

Change-Id: I73b634a33f2c03d2392cf1a9ff419a69bbbddb44