Multiple fixes for Omnigroupmember.load
It seems that a big reason why phone consent pulls were not working is
that we were expecting these to be phone-only records. But, the form
has been adjusted to retrieve emails as well. Given this we
can't assume that the opt in date is relevant and
we need to do an extra lookup to fetch the consent down.

This means we are determining the opt-in-i-ness in the first place
from the Acoustic group criteria - which means we need a parameter to
tell the code if it is an opt-in (or opt-out) group.

There are also fixes to use a parent class & to fix the opt_in
on that class to NULL for not set to prevent being treated as 0

Locally I have been testing with
 wmf-cv api4 Omnigroupmember.load groupIdentifier=44718026 jobIdentifier=consent_44718026 isConsentOptInGroup=1 limit=2 debug=1

Note the additional parameter isConsentOptInGroup=1 needs to be added to process control - it is
information backed into the Acoustic group. In addition we should
add to process-control the group to add the contact to - this is only
applicable when the email is supplied as well as the phone - in which case it
is like any other rml in terms of adding the contact and adding to a group

Bug: T379702
Change-Id: Ie96740e244985cae32d2699b0151efaec586bd30