Send recurring failure email via coworker
Get it out of the charge loop, potentially get better error visibility.

Re-uses the civicrm_api4_queue function added for snooze in the omnimail
extension, and cribs the enqueuing code from

  https://phabricator.wikimedia.org/diffusion/WFCG/browse/master/ext/org.wikimedia.omnimail/Civi/Api4/Action/Omnicontact/Snooze.php$40

Note that this patch does not affect the failure emails sent via the
recurring queue consumer, just the ones in the nightly charge job.

Bug: T356073
Change-Id: I14fde877fa875fca51b8c931e5ba5cd88293e924