Add api to remove contacts from master suppression list
This adds an api to remove a contact from an Acoustic list.

In Acoustic land the master suppression list is a list or group in the
same way an actual list or group is.

Further our entire database is a list or group - so this api can also be used to remove
from a 'normal' group or to delete from the database.

The basic method is to pass in email and (acoustic) group_identifier.

However, if you pass in email it will look it up. Further it will look
up group_identifier from a passed in groupID. Note this relies on the custom
field attached to the group being populated with the Acoustic ID - ie
Group_Metadata.remote_group_identifier

I filled in this group identifier on prod on the master suppression with the remote master
suppression list ID and cleared out all the others (old tests).

This will allow us to remove the opted in ones off the suppression list

I created this search kit of them
https://civicrm.wikimedia.org/civicrm/admin/search#/edit/6501 & based on my brief
digging into those contacts I found a lot seemed to be mergeable with otherwise
opted out contacts - so it might be worth seeing if donor relations can do a dedupe
pass on them first.

I think we should look at adding a hook to fire this whenever a group contact
record is updated to 'Removed' - the hook would look up whether
the group is linked to an Acoustic group (only master suppression list at the moment)
and if so queue a group removal.

I originally thought this would cause problems with our nightly export but on reviewing
the Suppression group membership is sticky at the Acoustic end but in our silverpop
database it constantly removes all contacts from our local
suppression list table if they are opted in and not opted out

wmf-cv api4 Omnigroupmember.delete email=mouse@example.org groupID=9574332
wmf-cv api4 Omnigroupmember.delete contactID=2815  groupID=9574332
wmf-cv api4 Omnigroupmember.delete contactID=2815  groupID=2434
wmf-cv api4 Omnigroupmember.delete contactID=mouse@example.org  groupID=2434

Bug: T387167
Change-Id: I6baca52be58f713e807ed5a6db3f8d0db6348369