CiviCRM upgrade to 6.8.0 rc
Note that this is our second upgrade since switching to standalone so not quite bedded in
(our prior process was harder). What I have done is

1) checked out rc branch 6.8 locally & committed it
 - I'm relying on our gitignore to keep out the files we don't want - e.g tests
2) verified vendor updates - I checked the remove composer.lock and the
only change is to composer.json - Drop support for older firebase/php-jwt. Require 6+
Given that I'm just gonna try a composer update as a follow on (I think we might not
need / get that update)
3) ran composer install in core folder & then deleted vendor again. I did this
because although we have incorporated civicrm/vendor in our main vendor using
composer-merge-plugin it DOES apply a couple of extension patches. These rarely change
but if they had I think running composer install in the core directory would have shown
that up.

So,ultimately this commit is exactly what is in the core 6.8 branch less a few ignored
files from our git ignore

Bug: T406645
Change-Id: I840c6aa3599c5e70feb5f011989c48563185ecd7