Record fees when they come in
This puts them against the anonymous donor (which is already
excluded from wmf_donor triggers). If it ever becomes a thing we
could move to specific-donors.

Also at this stage there is no specific financial type
or payment instrument type - we can also iterate on that as
I work through it wil stakeholders. Easy enough to move a handful

Bug: T152249
Change-Id: I6c2bcd5bfb1fe83eb54a4785b979ff7ce7908da9