Add settled fields spec provider
This allows us to display in search kit the batch values that are in the given batch.

Our current search allows us to find all contributions that are in the batch of that have a refund
in that batch. However, in the latter case the original amount of the contribution is
currently displayed in the batch when it actually belongs to an earlier batch,
giving unbalanced totals

https://civicrm.wikimedia.org/civicrm/contribution/settled#?contribution_settlement.settlement_batch_reference,contribution_settlement.settlement_batch_reversal_reference=paypal_20260129_CAD

Bug: T412172

Change-Id: I23e3a2b82e99fc0ef91a9234be48398de49b3bde